Make the regridding method a class method of the preprocessor
