+++ tests/preprocess/test_utils.py
@@ -1,9 +1,8 @@
 import xarray as xr
 import numpy as np
-import pytest
 
 from src.utils import Region
-from src.preprocess.utils import select_bounding_box, regrid
+from src.preprocess.utils import select_bounding_box
 
 
 def _make_dataset(size):
@@ -58,33 +57,3 @@ class TestSelectBoundingBox:
             f'Expected output subset to have size (50, 100), got {subset.VHI.values.shape}'
         assert max(subset.lon.values) < 0, \
             f'Got a longitude greater than 0, {max(subset.lon.values)}'
-
-
-class TestRegridding:
-
-    def test_regridding(self):
-
-        size_reference = (100, 100)
-        size_target = (1000, 1000)
-
-        reference_ds, _, _ = _make_dataset(size_reference)
-        target_ds, _, _ = _make_dataset(size_target)
-
-        regridded_ds = regrid(target_ds, reference_ds)
-
-        assert regridded_ds.VHI.values.shape == size_reference, \
-            f'Expected regridded Dataset to have shape {size_reference}, ' \
-            f'got {regridded_ds.VHI.values.shape}'
-
-    def test_incorrect_method(self):
-        size_reference = (100, 100)
-        size_target = (1000, 1000)
-
-        reference_ds, _, _ = _make_dataset(size_reference)
-        target_ds, _, _ = _make_dataset(size_target)
-
-        with pytest.raises(AssertionError) as e:
-            regrid(target_ds, reference_ds, method='woops!')
-        expected_message_contains = 'not an acceptable regridding method. Must be one of'
-        assert expected_message_contains in str(e), \
-            f'Expected {e} to contain {expected_message_contains}'
